fix setting last pipeline and rename to repo-card
